\name{MPermutation-package}
\alias{MPermutation-package}
\alias{MPermutation}
\docType{package}
\title{
Multiset Permutation
}
\description{
Label switching algorithm, allow multiple groups.
}
\details{
\tabular{ll}{
Package: \tab MPermutation\cr
Type: \tab Package\cr
Version: \tab 1.13\cr
Date: \tab 2007-09-24\cr
License: \tab GPL\cr
}

}
\author{
Kun Liang

Maintainer: liangkun1@gmail.com

}

\examples{
MPfile(c(2,3))
MPmatrix(c(2,3))
}
