\name{MPmatrix}
\alias{MPmatrix}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Multiset Permutation, return in Matrix }
\description{
  Label switching algorithm, allow multiple groups.
}
\usage{
MPmatrix(vec)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vec}{ A vector of integer as group sizes }
}

\value{
  A matrix.
}

\author{ Kun Liang }

\examples{
MPmatrix(c(2,3))
}
